refactor into plasma tests
