+++ projects/tests/weibel/gam10.ini
@@ -1,6 +1,6 @@
 
 [io]
-outdir: "gam10"
+outdir: "weibel/gam10"
 interval: 10       #output frequency in units of simulation steps for analysis files
 full_interval: -1 #output frequency to write full simulation snapshots
 restart:  1000    #frequency to write restart files (these overwrite previous files)
