+++ projects/tests/weibel/gam100.ini
@@ -1,6 +1,6 @@
 
 [io]
-outdir: "gam50"
+outdir: "weibel/gam100"
 interval: 10       #output frequency in units of simulation steps for analysis files
 full_interval: -1 #output frequency to write full simulation snapshots
 restart:  1000    #frequency to write restart files (these overwrite previous files)
@@ -19,7 +19,7 @@ Nspecies: 2        #number of species (typically 2)
 delgam:  1.0e-5    #temperature
 temp_ratio: 1.0    #T_i/T_e
 
-gamma: 50.0      #bulk flow speed
+gamma: 100.0      #bulk flow speed
 
 me: -1.0        #electron mass-to-charge
 mi: -1.0        #ion mass-to-charge
